backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
numpy==1.26.4
psycopg2-binary==2.9.9
pymongo==4.7.2
cloudinary==1.40.0
python-dotenv==1.0.1
